feat: Replace scanning gyms with only query by country
